Extract AdventDay class.
